Separate follow feature into independent service
